requirements.txt
streamlit>=1.31.0
python-docx>=1.1.0
openpyxl>=3.1.2
google-generativeai>=0.3.0
python-dotenv>=1.0.0
